Move Buffer to util directory.
