+++ opts/iapd.go
@@ -5,6 +5,7 @@ import (
 	"time"
 
 	"github.com/mdlayher/dhcp6"
+	"github.com/mdlayher/dhcp6/util"
 )
 
 // IAPD represents an Identity Association for Prefix Delegation, as
@@ -54,7 +55,7 @@ func (i *IAPD) MarshalBinary() ([]byte, error) {
 	// 4 bytes: T1
 	// 4 bytes: T2
 	// N bytes: options slice byte count
-	buf := dhcp6.NewBuffer(nil)
+	buf := util.NewBuffer(nil)
 
 	buf.WriteBytes(i.IAID[:])
 	buf.Write32(uint32(i.T1 / time.Second))
@@ -70,7 +71,7 @@ func (i *IAPD) MarshalBinary() ([]byte, error) {
 // io.ErrUnexpectedEOF is returned.
 func (i *IAPD) UnmarshalBinary(b []byte) error {
 	// IAPD must contain at least an IAID, T1, and T2.
-	buf := dhcp6.NewBuffer(b)
+	buf := util.NewBuffer(b)
 	if buf.Len() < 12 {
 		return io.ErrUnexpectedEOF
 	}
