+++ opts/relaymessage.go
@@ -5,6 +5,7 @@ import (
 	"net"
 
 	"github.com/mdlayher/dhcp6"
+	"github.com/mdlayher/dhcp6/util"
 )
 
 // RelayMessage represents a raw RelayMessage generated by DHCPv6 relay agent, using RFC 3315,
@@ -43,7 +44,7 @@ func (rm *RelayMessage) MarshalBinary() ([]byte, error) {
 	// 16 bytes: link-address
 	// 16 bytes: peer-address
 	// N bytes: options slice byte count
-	b := dhcp6.NewBuffer(nil)
+	b := util.NewBuffer(nil)
 
 	b.Write8(uint8(rm.MessageType))
 	b.Write8(rm.HopCount)
@@ -59,7 +60,7 @@ func (rm *RelayMessage) MarshalBinary() ([]byte, error) {
 // If the byte slice does not contain enough data to form a valid RelayMessage,
 // ErrInvalidPacket is returned.
 func (rm *RelayMessage) UnmarshalBinary(p []byte) error {
-	b := dhcp6.NewBuffer(p)
+	b := util.NewBuffer(p)
 	// RelayMessage must contain at least message type, hop-count, link-address and peer-address
 	if b.Len() < 34 {
 		return io.ErrUnexpectedEOF
