+++ opts/statuscode.go
@@ -4,6 +4,7 @@ import (
 	"io"
 
 	"github.com/mdlayher/dhcp6"
+	"github.com/mdlayher/dhcp6/util"
 )
 
 // StatusCode represents a Status Code, as defined in RFC 3315, Section 5.4.
@@ -33,7 +34,7 @@ func NewStatusCode(code dhcp6.Status, message string) *StatusCode {
 func (s *StatusCode) MarshalBinary() ([]byte, error) {
 	// 2 bytes: status code
 	// N bytes: message
-	b := dhcp6.NewBuffer(nil)
+	b := util.NewBuffer(nil)
 	b.Write16(uint16(s.Code))
 	b.WriteBytes([]byte(s.Message))
 	return b.Data(), nil
@@ -44,7 +45,7 @@ func (s *StatusCode) MarshalBinary() ([]byte, error) {
 // If the byte slice does not contain enough data to form a valid StatusCode,
 // errInvalidStatusCode is returned.
 func (s *StatusCode) UnmarshalBinary(p []byte) error {
-	b := dhcp6.NewBuffer(p)
+	b := util.NewBuffer(p)
 	// Too short to contain valid StatusCode
 	if b.Len() < 2 {
 		return io.ErrUnexpectedEOF
