+++ opts/vendoropts.go
@@ -4,6 +4,7 @@ import (
 	"io"
 
 	"github.com/mdlayher/dhcp6"
+	"github.com/mdlayher/dhcp6/util"
 )
 
 // A VendorOpts is used by clients and servers to exchange
@@ -23,7 +24,7 @@ type VendorOpts struct {
 func (v *VendorOpts) MarshalBinary() ([]byte, error) {
 	// 4 bytes: EnterpriseNumber
 	// N bytes: options slice byte count
-	b := dhcp6.NewBuffer(nil)
+	b := util.NewBuffer(nil)
 	b.Write32(v.EnterpriseNumber)
 	v.Options.Marshal(b)
 
@@ -35,7 +36,7 @@ func (v *VendorOpts) MarshalBinary() ([]byte, error) {
 // VendorOpts, io.ErrUnexpectedEOF is returned.
 // If option-data are invalid, then ErrInvalidPacket is returned.
 func (v *VendorOpts) UnmarshalBinary(p []byte) error {
-	b := dhcp6.NewBuffer(p)
+	b := util.NewBuffer(p)
 	// Too short to be valid VendorOpts
 	if b.Len() < 4 {
 		return io.ErrUnexpectedEOF
