+++ util/buffer.go
@@ -1,4 +1,4 @@
-package dhcp6
+package util
 
 import (
 	"encoding/binary"
